server_enc/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libnum==1.7.1
MarkupSafe==2.1.2
pycryptodome==3.17
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
